phy: move simulation related utilities to sim_utils.py
