+++ litedram/phy/lpddr4/simsoc.py
@@ -10,7 +10,6 @@ import argparse
 from migen import *
 
 from litex.build.generic_platform import Pins, Subsignal
-from litex.build.sim import SimPlatform
 from litex.build.sim.config import SimConfig
 
 from litex.soc.interconnect.csr import CSR
@@ -26,21 +25,12 @@ from litedram.phy.model import DFITimingsChecker, _speedgrade_timings, _technolo
 from litedram.phy.lpddr4.simphy import LPDDR4SimPHY, DoubleRateLPDDR4SimPHY
 from litedram.phy.lpddr4.sim import LPDDR4Sim
 
+from litedram.phy.sim_utils import Clocks, CRG, Platform
+
 # Platform -----------------------------------------------------------------------------------------
 
 _io = [
-    # clocks added later
-    ("sys_rst", 0, Pins(1)),
-
-    ("serial", 0,
-        Subsignal("source_valid", Pins(1)),
-        Subsignal("source_ready", Pins(1)),
-        Subsignal("source_data",  Pins(8)),
-        Subsignal("sink_valid",   Pins(1)),
-        Subsignal("sink_ready",   Pins(1)),
-        Subsignal("sink_data",    Pins(8)),
-    ),
-
+    # clocks added in main()
     ("lpddr4", 0,
         Subsignal("clk",   Pins(1)),
         # Subsignal("clk_n",   Pins(1)),
@@ -56,43 +46,7 @@ _io = [
     ),
 ]
 
-class Platform(SimPlatform):
-    def __init__(self):
-        SimPlatform.__init__(self, "SIM", _io)
-
 # Clocks -------------------------------------------------------------------------------------------
-
-class Clocks(dict):  # FORMAT: {name: {"freq_hz": _, "phase_deg": _}, ...}
-    def names(self):
-        return list(self.keys())
-
-    def add_io(self, io):
-        for name in self.names():
-            io.append((name + "_clk", 0, Pins(1)))
-
-    def add_clockers(self, sim_config):
-        for name, desc in self.items():
-            sim_config.add_clocker(name + "_clk", **desc)
-
-class _CRG(Module):
-    def __init__(self, platform, domains=None):
-        if domains is None:
-            domains = ["sys"]
-        # request() before creating domains to avoid signal renaming problem
-        domains = {name: platform.request(name + "_clk") for name in domains}
-
-        self.clock_domains.cd_por = ClockDomain(reset_less=True)
-        for name in domains.keys():
-            setattr(self.clock_domains, "cd_" + name, ClockDomain(name=name))
-
-        int_rst = Signal(reset=1)
-        self.sync.por += int_rst.eq(0)
-        self.comb += self.cd_por.clk.eq(self.cd_sys.clk)
-
-        for name, clk in domains.items():
-            cd = getattr(self, "cd_" + name)
-            self.comb += cd.clk.eq(clk)
-            self.comb += cd.rst.eq(int_rst)
 
 def get_clocks(sys_clk_freq):
     return Clocks({
@@ -115,7 +69,7 @@ class SimSoC(SoCCore):
     def __init__(self, clocks, log_level,
             auto_precharge=False, with_refresh=True, trace_reset=0, disable_delay=False,
             masked_write=True, double_rate_phy=False, finish_after_memtest=False, **kwargs):
-        platform     = Platform()
+        platform     = Platform(_io, clocks)
         sys_clk_freq = clocks["sys"]["freq_hz"]
 
         # SoCCore ----------------------------------------------------------------------------------
@@ -127,7 +81,7 @@ class SimSoC(SoCCore):
             **kwargs)
 
         # CRG --------------------------------------------------------------------------------------
-        self.submodules.crg = _CRG(platform, clocks.names())
+        self.submodules.crg = CRG(platform, clocks)
 
         # Debugging --------------------------------------------------------------------------------
         platform.add_debug(self, reset=trace_reset)
@@ -341,7 +295,6 @@ def main():
     sim_config = SimConfig()
     sys_clk_freq = int(float(args.sys_clk_freq))
     clocks = get_clocks(sys_clk_freq)
-    clocks.add_io(_io)
     clocks.add_clockers(sim_config)
 
     # Configuration --------------------------------------------------------------------------------
